refactor 2015 talk transcript links
The 2015 talk transcripts are named after the author slugs, but the 2014
talks do not. This leads to a minor difference in the construction of
the transcript links. To support extracting the recording page
JavaScript into a separate, shared file, update the 2015 talk data to
include the (partial) filename of the transcript.